# Speed up `degree(::GAPGroupClassFunction)`