Resolves https://github.com/oscar-system/Oscar.jl/issues/5112.

With this change, only the actually used entries of the character are converted from GAP to Oscar, instead of all of them.

Speed comparison:
```julia
julia> T = character_table(small_group(512, 1));

julia> @time degree(chi);
  0.258867 seconds (26.98 k allocations: 22.903 MiB, 89.76% gc time) # master
  0.000072 seconds (17 allocations: 752 bytes) # this PR

julia> @time map(degree, T);
  7.928418 seconds (13.34 M allocations: 7.742 GiB, 5.83% gc time) # master
  0.001854 seconds (9.22 k allocations: 388.109 KiB) # this PR
```

cc @thofma